requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
google-auth==2.6.2
requests==2.27.1
